backend/requirements.txt
fastapi
uvicorn
pandas
duckdb
langchain
langgraph
requests
python-dotenv
openai
